Finish transition to ArrowheadServer
Prior to commit 44c032d ArrowheadSystem was an abstract class that was
extended by concrete classes that provide their own application-level
transport protocols. It was realized, however, that individuals systems
should be able to expose services with distinct application and
transport-level protocols. For this reason, ArrowheadSystem has been
made into a regular class, and the work of handling services for
concrete transports has been delegated to instances of the
ArrowheadServer class. ArrowheadServers are instantiated as required as
new services are added to an ArrowheadSystem, effectively hiding them
from Kalix users.