Try to fix DTO type resolution Heisenbug
It seemed the bug was caused by the JDK returning the wrong type
when multiple types were around with the same names (albeit being
in different packages). I simply try to resolve the type again if
type resolution fails. I'm not sure how effective this is going
to be in practice.

The bug might be related to the version of Maven I used when the
bug happened. I can't seem to reproduce it with Gradle, at least.