Make Future#await() throw if called from Netty thread
This is an attempt to prevent issues such as #19 from every happening
again. Completely banning blocking of Netty threads might be a bit of
a long shot, but I've added an admonition to open a discussion about
blocking in the thrown exception. If a justifiable use of blocking
Netty threads would arise, we can cater for that then.